refactor(icons): move the icons from temp to projectV2 folder
