+++ public/assets/images/icons/projectV2/DownloadIcon.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IconProps } from '../../features/common/types/common';
+import type { IconProps } from '../../../../../src/features/common/types/common';
 
 const DownloadIcon = ({ width, color }: IconProps) => {
   return (
