+++ public/assets/images/icons/projectV2/ExpandIcon.tsx
